refactor: 提取重复代码为独立 trait 和工具方法
- 提取中间件 shouldPassThrough 公共逻辑到 Helper::shouldPassThrough()
- 提取 Select model 选项逻辑到 HasModelOptions trait
- 提取控制器菜单树字段逻辑到 HasMenuTreeField trait
- 提取验证属性格式化逻辑到 HasFormatValidationAttribute trait
- 消除中间件间的相互调用，保持高内聚低耦合
