+++ src/Form/Field/HasFormatValidationAttribute.php
@@ -1,0 +1,44 @@
+<?php
+
+namespace Dcat\Admin\Form\Field;
+
+use Illuminate\Support\Arr;
+use Illuminate\Support\Str;
+
+trait HasFormatValidationAttribute
+{
+    /**
+     * Format validation attribute.
+     *
+     * @param  array  $input
+     * @param  string  $label
+     * @param  string|array  $column
+     * @return array
+     */
+    protected function formatValidationAttribute($input, $label, $column)
+    {
+        $new = $attributes = [];
+
+        if (is_array($column)) {
+            foreach ($column as $index => $col) {
+                $new[$col.$index] = $col;
+            }
+        }
+
+        foreach (array_keys(Arr::dot($input)) as $key) {
+            if (is_string($column)) {
+                if (Str::endsWith($key, ".$column")) {
+                    $attributes[$key] = $label;
+                }
+            } else {
+                foreach ($new as $k => $val) {
+                    if (Str::endsWith($key, ".$k")) {
+                        $attributes[$key] = $label."[$val]";
+                    }
+                }
+            }
+        }
+
+        return $attributes;
+    }
+}
