+++ src/Form/Field/Select.php
@@ -2,16 +2,14 @@
 
 namespace Dcat\Admin\Form\Field;
 
-use Dcat\Admin\Exception\RuntimeException;
 use Dcat\Admin\Form\Field;
-use Dcat\Admin\Support\Helper;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Support\Arr;
+use Dcat\Admin\Traits\HasModelOptions;
 
 class Select extends Field
 {
     use CanCascadeFields;
     use CanLoadFields;
+    use HasModelOptions;
     use Sizeable;
 
     protected $cascadeEvent = 'change';
@@ -79,45 +77,6 @@ class Select extends Field
     public function groups(array $groups)
     {
         $this->groups = $groups;
-
-        return $this;
-    }
-
-    /**
-     * Load options from current selected resource(s).
-     *
-     * @param  string  $model
-     * @param  string  $idField
-     * @param  string  $textField
-     * @return $this
-     */
-    public function model($model, string $idField = 'id', string $textField = 'name')
-    {
-        if (! class_exists($model)
-            || ! in_array(Model::class, class_parents($model))
-        ) {
-            throw new RuntimeException("[$model] must be a valid model class");
-        }
-
-        $this->options = function ($value) use ($model, $idField, $textField) {
-            if (empty($value)) {
-                return [];
-            }
-
-            $resources = [];
-
-            if (is_array($value)) {
-                if (Arr::isAssoc($value)) {
-                    $resources[] = Arr::get($value, $idField);
-                } else {
-                    $resources = array_column($value, $idField);
-                }
-            } else {
-                $resources[] = $value;
-            }
-
-            return $model::whereIn($idField, $resources)->pluck($textField, $idField)->toArray();
-        };
 
         return $this;
     }
