+++ src/Http/Controllers/PermissionController.php
@@ -11,6 +11,7 @@ use Illuminate\Support\Str;
 
 class PermissionController extends AdminController
 {
+    use HasMenuTreeField;
     protected function title()
     {
         return trans('admin.permissions');
@@ -116,21 +117,7 @@ class PermissionController extends AdminController
                 ->options($this->getRoutes());
 
             if ($bindMenu) {
-                $form->tree('menus', trans('admin.menu'))
-                    ->treeState(false)
-                    ->setTitleColumn('title')
-                    ->nodes(function () {
-                        $model = config('admin.database.menu_model');
-
-                        return (new $model())->allNodes();
-                    })
-                    ->customFormat(function ($v) {
-                        if (! $v) {
-                            return [];
-                        }
-
-                        return array_column($v, 'id');
-                    });
+                $this->addMenuTreeField($form);
             }
 
             $form->display('created_at', trans('admin.created_at'));
@@ -139,8 +126,7 @@ class PermissionController extends AdminController
             $form->disableViewButton();
             $form->disableViewCheck();
         })->saved(function () {
-            $model = config('admin.database.menu_model');
-            (new $model())->flushCache();
+            $this->flushMenuCache();
         });
     }
