+++ src/Traits/HasModelOptions.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace Dcat\Admin\Traits;
+
+use Dcat\Admin\Exception\RuntimeException;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Arr;
+
+trait HasModelOptions
+{
+    /**
+     * Load options from current selected resource(s).
+     *
+     * @param  string  $model
+     * @param  string  $idField
+     * @param  string  $textField
+     * @return $this
+     */
+    public function model($model, string $idField = 'id', string $textField = 'name')
+    {
+        if (! class_exists($model)
+            || ! in_array(Model::class, class_parents($model))
+        ) {
+            throw new RuntimeException("[$model] must be a valid model class");
+        }
+
+        $this->options = function ($value) use ($model, $idField, $textField) {
+            if (empty($value)) {
+                return [];
+            }
+
+            $resources = [];
+
+            if (is_array($value)) {
+                if (Arr::isAssoc($value)) {
+                    $resources[] = Arr::get($value, $idField);
+                } else {
+                    $resources = array_column($value, $idField);
+                }
+            } else {
+                $resources[] = $value;
+            }
+
+            return $model::whereIn($idField, $resources)->pluck($textField, $idField)->toArray();
+        };
+
+        return $this;
+    }
+}
